requirements.txt
numpy
scipy
pycollada  # to load .dae files
urdfdom-py @ git+https://github.com/ros/urdf_parser_py.git@1.2.1
pyyaml  # dependency for urdf_parser_py (see package.xml)
lxml  # dependency for urdf_parser_py (see package.xml)
trimesh
rerun-sdk>=0.12
